requirements.txt
annotated-types==0.7.0
anyio==4.10.0
beautifulsoup4==4.13.5
blinker==1.9.0
bs4==0.0.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.3.0
faiss-cpu==1.12.0
filelock==3.19.1
Flask==3.1.2
fsspec==2025.7.0
groq==0.31.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
lxml==6.0.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
PyYAML==6.0.2
regex==2025.8.29
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.15.3
sentence-transformers==5.1.0
sniffio==1.3.1
soupsieve==2.8
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.0
torch==2.8.0
tqdm==4.67.1
transformers==4.56.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3